execconnect/requirements.txt
# ExecConnect consumes approved trades, executes them (mock or real) via RabbitMQ messages, logs execution in DB or optional broker APIs

pika==1.3.1
psycopg2-binary==2.9.7
requests==2.28.1